Refactoring styles, components and functions
