+++ client/components/object-photos-table/ObjectPhotoTable.tsx
@@ -1,7 +1,7 @@
 import { ApiModel } from '@/api'
-import { formatSecondsToExposure } from '@/functions/helpers'
 import { getFilterColor } from '@/tools/colors'
 import { formatDate } from '@/tools/dates'
+import { formatSecondsToExposure } from '@/tools/helpers'
 import { createPhotoTitle, createSmallPhotoUrl } from '@/tools/photos'
 import { useTranslation } from 'next-i18next'
 import Image from 'next/image'
