+++ client/components/weather/styles.module.sass
@@ -1,8 +1,12 @@
 .weather
     margin-bottom: 10px
 
-    .blockTitle
-        margin: 0
+    .toolbar
+        display: flex
+        justify-content: space-between
+
+        .title
+            margin: 0
 
     .update
         margin: 2px 0 5px 0
