+++ client/styles/globals.sass
@@ -159,7 +159,7 @@ blockquote
 
 .errorPage
     text-align: center
-    padding: 50px 0 100px
+    padding: 100px 0
 
     $background404Image: '/images/error404_background.jpg'
